feat(datastream): proto stream read/write (#347)
* feat(datastream): proto stream write

* feat(datastream): read proto stream

* fixing up some stream related things post rebase

* updating to latest proto for data stream

* some fixes for syncing on the new proto format

* only recover the data stream if execution has done any work

* simplify looking up fork id by batch

* fixing up datastream write

* fixing up tests related to data stream proto changes

* post merge fixes from zkevm branch

* more progress on proto datastream

* hack file for hosting a local datastream without erigon attached

* keep track of the fork id within the stream client rather than the batches stage

* handle ger updates in the new proto format for writing the stream

* fix stream related tests

* remove accidentally committed debug files

* datastream related code tidy up

* fix post zkevm merge problems in proto datastream

* fix(witness): closed growing number of cursors (#416)

* add more detail to proto datastream

* add batch number to batch end proto

* add more details for proto format datastream

* fix(merge): hermez db merge fix

* fix for updating chain config with fork id changes

* fixing up tests after changing the fork lookup code for chain config

* do not write the batch end during genesis block write

this will be handled by the writing of block 1 (injected batch) once it detects that the batch number has changed h

* fix bad build on proto branch

* feat(datastream): l1 cache, debug fix

* feat(docs): cache headers, documentation

---------

+++ zk/stages/stage_sequence_execute_transactions.go
@@ -39,7 +39,7 @@ LOOP:
 		}
 		if err := cfg.txPoolDb.View(context.Background(), func(poolTx kv.Tx) error {
 			slots := types2.TxsRlp{}
-			_, count, err = cfg.txPool.YieldBest(yieldSize, &slots, poolTx, executionAt, getGasLimit(uint16(forkId)), alreadyYielded)
+			_, count, err = cfg.txPool.YieldBest(yieldSize, &slots, poolTx, executionAt, getGasLimit(forkId), alreadyYielded)
 			if err != nil {
 				return err
 			}
