+++ zk/tx/tx_test.go
@@ -68,7 +68,7 @@ func TestDecodePre155Tx(t *testing.T) {
 	pre155 := "0xf86780843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77c1ba0b7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feba0246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed"
 	pre155Bytes, err := hex.DecodeString(pre155[2:])
 	require.NoError(t, err)
-	tx, _, err := DecodeTx(pre155Bytes, 0, uint16(constants.ForkID5Dragonfruit))
+	tx, _, err := DecodeTx(pre155Bytes, 0, uint64(constants.ForkID5Dragonfruit))
 	require.NoError(t, err)
 	v, r, s := tx.RawSignatureValues()
 	assert.Equal(t, "0x1275fbb540c8efC58b812ba83B0D0B8b9917AE98", tx.GetTo().String())
@@ -85,7 +85,7 @@ func TestDecodePost155Tx(t *testing.T) {
 	post155 := "0xf86780843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77c1ba0b7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feba0246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed"
 	post155Bytes, err := hex.DecodeString(post155[2:])
 	require.NoError(t, err)
-	tx, pct, err := DecodeTx(post155Bytes, 75, uint16(constants.ForkID5Dragonfruit))
+	tx, pct, err := DecodeTx(post155Bytes, 75, uint64(constants.ForkID5Dragonfruit))
 	require.NoError(t, err)
 	v, r, s := tx.RawSignatureValues()
 	assert.Equal(t, "0x1275fbb540c8efC58b812ba83B0D0B8b9917AE98", tx.GetTo().String())
